Assignment 1
Assignment 1 for the Advanced Statistical Inference subject (done in the first semester of the 2023-2024 course). This assignment was for the first unit of the subject, the contents of this unit are: Preliminaries and Inferential questions, The frequentist approach: point estimation, some finite sample properties. The frequentist approach: Hypothesis testing and permutational approach. Interval estimation. The Bayesian approach: prior and posterior densities, marginal likelihood. Appendix: common tests.