Files to fit a Bayesian frailty model
These files provide the code to fit the Bayesian frailty model presented in the memory (also provided in this repository). There is the Stan model in "frail_def.stan", the R code running this Stan model in "stan_run.R", and an Rmarkdown where several plots using the results of the Stan model are done.